requirements.txt
flask
flask-cors
faiss-cpu
numpy
python-dotenv
google-generativeai
sentence-transformers
python-docx